requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
openai>=1.12.0
notebook>=7.0.0
ipykernel>=6.0.0
langsmith>=0.0.87
langserve>=0.0.41
beautifulsoup4>=4.12.2
pypdf>=3.15.1
wikipedia>=1.4.0
arxiv>=1.4.7
pymupdf>=1.23.8 